PowerManagerService: Fix arm64 build
Add static_cast<int64_t> for 64bit compatiblity

+++ services/core/jni/com_android_server_power_PowerManagerService.cpp
@@ -166,7 +166,7 @@ static void nativeCpuBoost(JNIEnv *env, jobject clazz, jint duration) {
 static void nativeSetPowerProfile(JNIEnv *env, jobject clazz, jint profile) {
     // Tell the Power HAL to select a power profile
     if (gPowerModule && gPowerModule->powerHint) {
-        gPowerModule->powerHint(gPowerModule, POWER_HINT_SET_PROFILE, (void *) profile);
+        gPowerModule->powerHint(gPowerModule, POWER_HINT_SET_PROFILE, (void *)(static_cast<int64_t>(profile)));
     }
 }
